Move back-end methods from Products class to appropriate classes in admin section
