+++ wp-express-checkout/includes/class-products.php
@@ -2,32 +2,11 @@
 
 namespace WP_Express_Checkout;
 
-use WP_Express_Checkout\Admin\Products_Meta_Boxes;
-
 class Products {
 
 	static $products_slug = "ppec-products";
 
-	protected static $instance = null;
-
-	function __construct() {
-		if ( is_admin() ) {
-			// products meta boxes handler.
-			new Products_Meta_Boxes();
-		}
-	}
-
-	public static function get_instance() {
-
-		// If the single instance hasn't been set, set it now.
-		if ( null === self::$instance ) {
-			self::$instance = new self();
-		}
-
-		return self::$instance;
-	}
-
-	function register_post_type() {
+	public static function register_post_type() {
 
 		// Products post type.
 		$labels = array(
@@ -69,108 +48,6 @@ class Products {
 		);
 
 		register_post_type( self::$products_slug, $args );
-
-		// add custom columns for list view.
-		add_filter( 'manage_' . self::$products_slug . '_posts_columns', array( $this, 'manage_columns' ) );
-		add_action( 'manage_' . self::$products_slug . '_posts_custom_column', array( $this, 'manage_custom_columns' ), 10, 2 );
-		// set custom columns sortable.
-		add_filter( 'manage_edit-' . self::$products_slug . '_sortable_columns', array( $this, 'manage_sortable_columns' ) );
-		// set custom messages on post save\update etc.
-		add_filter( 'post_updated_messages', array( $this, 'post_updated_messages' ) );
-	}
-
-	function post_updated_messages( $messages ) {
-		$post      = get_post();
-		$post_type = get_post_type( $post );
-		$slug      = self::$products_slug;
-
-		if ( $post_type === self::$products_slug ) {
-			$permalink = get_permalink( $post->ID );
-			$view_link = sprintf( ' <a href="%s">%s</a>', esc_url( $permalink ), __( 'View product', 'wp-express-checkout' ) );
-
-			$preview_permalink = add_query_arg( 'preview', 'true', $permalink );
-			$preview_link      = sprintf( ' <a target="_blank" href="%s">%s</a>', esc_url( $preview_permalink ), __( 'Preview product', 'wp-express-checkout' ) );
-
-			$messages[ $slug ]     = $messages['post'];
-			$messages[ $slug ][1]  = __( 'Product updated.', 'wp-express-checkout' ) . $view_link;
-			$messages[ $slug ][4]  = __( 'Product updated.', 'wp-express-checkout' );
-			$messages[ $slug ][6]  = __( 'Product published.', 'wp-express-checkout' ) . $view_link;
-			$messages[ $slug ][7]  = __( 'Product saved.', 'wp-express-checkout' );
-			$messages[ $slug ][8]  = __( 'Product submitted.', 'wp-express-checkout' ) . $preview_link;
-			$messages[ $slug ][10] = __( 'Product draft updated.', 'wp-express-checkout' ) . $preview_link;
-		}
-		return $messages;
-	}
-
-	function manage_columns( $columns ) {
-		unset( $columns );
-		$columns = array(
-			'cb'        => '<input type="checkbox">',
-			'thumbnail' => __( 'Thumbnail', 'wp-express-checkout' ),
-			'title'     => __( 'Product Name', 'wp-express-checkout' ),
-			'id'        => __( 'ID', 'wp-express-checkout' ),
-			'price'     => __( 'Price', 'wp-express-checkout' ),
-			'shortcode' => __( 'Shortcode', 'wp-express-checkout' ),
-			'date'      => __( 'Date', 'wp-express-checkout' ),
-		);
-		return $columns;
-	}
-
-	function manage_custom_columns( $column, $post_id ) {
-		$main = Main::get_instance();
-		switch ( $column ) {
-			case 'id':
-				echo $post_id;
-				break;
-			case 'thumbnail':
-				$thumb_url = get_post_meta( $post_id, 'wpec_product_thumbnail', true );
-				if ( ! $thumb_url ) {
-					$thumb_url = WPEC_PLUGIN_URL . '/assets/img/product-thumb-placeholder.png';
-				}
-				$edit_link = get_edit_post_link( $post_id );
-				$title     = __( 'Edit Product', 'wp-express-checkout' );
-				?>
-				<span class="wpec-product-thumbnail-container">
-					<a href="<?php echo esc_attr( $edit_link ); ?>" title="<?php echo $title; ?>">
-						<div style="padding: 50px 0; max-width: 100px; background-image: url(<?php echo esc_url( $thumb_url ); ?>); background-size: cover; background-position: center;"></div>
-					</a>
-				</span>
-				<?php
-				break;
-			case 'price':
-				$price_args = array_merge(
-					array(
-						'price'           => 0,
-						'shipping'        => 0,
-						'tax'             => 0,
-						'quantity'        => 1,
-					),
-					array(
-						'name'            => get_the_title( $post_id ),
-						'price'           => get_post_meta( $post_id, 'ppec_product_price', true ),
-						'shipping'        => ( '' === get_post_meta( $post_id, 'wpec_product_shipping', true ) ) ? $main->get_setting( 'shipping' ) : get_post_meta( $post_id, 'wpec_product_shipping', true ),
-						'tax'             => ( '' === get_post_meta( $post_id, 'wpec_product_tax', true ) ) ? $main->get_setting( 'tax' ) : get_post_meta( $post_id, 'wpec_product_tax', true ),
-						'quantity'        => get_post_meta( $post_id, 'ppec_product_quantity', true ),
-						'product_id'      => $post_id,
-					)
-				);
-				$wpec_shortcode = Shortcodes::get_instance();
-				$output = $wpec_shortcode->generate_price_tag( $price_args );
-				$output = apply_filters( 'ppec_products_table_price_column', $output, $price_args, $post_id );
-				echo $output;
-				break;
-			case 'shortcode':
-				?>
-				<input type="text" name="ppec_product_shortcode" class="ppec-select-on-click" readonly value="[wp_express_checkout product_id=&quot;<?php echo $post_id; ?>&quot;]">
-				<?php
-				break;
-		}
-	}
-
-	function manage_sortable_columns( $columns ) {
-		$columns['id']    = 'id';
-		$columns['price'] = 'price';
-		return $columns;
 	}
 
 }
