+++ wp-express-checkout/tests/admin/includes/test-class-products-list.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace WP_Express_Checkout\Admin;
+
+/**
+ * Generated by PHPUnit_SkeletonGenerator on 2021-06-28 at 07:52:57.
+ */
+class Products_ListTest extends \WP_UnitTestCase {
+
+	/**
+	 * @var Products
+	 */
+	protected $object;
+
+	/**
+	 * Sets up the fixture, for example, opens a network connection.
+	 * This method is called before a test is executed.
+	 */
+	public function setUp() {
+		$this->object = new Products_List;
+	}
+
+	/**
+	 * @covers WP_Express_Checkout\Products::manage_columns
+	 * @todo   Implement testManage_columns().
+	 */
+	public function testManage_columns() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+				'This test has not been implemented yet.'
+		);
+	}
+
+	/**
+	 * @covers WP_Express_Checkout\Products::manage_custom_columns
+	 * @todo   Implement testManage_custom_columns().
+	 */
+	public function testManage_custom_columns() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+				'This test has not been implemented yet.'
+		);
+	}
+
+	/**
+	 * @covers WP_Express_Checkout\Products::manage_sortable_columns
+	 * @todo   Implement testManage_sortable_columns().
+	 */
+	public function testManage_sortable_columns() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+				'This test has not been implemented yet.'
+		);
+	}
+
+}
