+++ wp-express-checkout/tests/includes/test-class-orders.php
@@ -26,15 +26,8 @@ class OrdersTest extends \WP_UnitTestCase {
 	public function testRegister_post_type() {
 		unregister_post_type( Orders::PTYPE );
 		$this->assertNull( get_post_type_object( Orders::PTYPE ) );
-		$this->object->register_post_type();
+		Orders::register_post_type();
 		$this->assertInstanceOf( 'WP_Post_Type', get_post_type_object( Orders::PTYPE ) );
-	}
-
-	/**
-	 * @covers WP_Express_Checkout\Orders::get_instance
-	 */
-	public function testGet_instance() {
-		$this->assertInstanceOf( 'WP_Express_Checkout\Orders', Orders::get_instance() );
 	}
 
 	/**
